Add TaskListStatus.empty (#212)
The adaptive TaskList scaler currently relies on the backlog_count_hint to determine whether a TaskList is empty and can be safely removed. This value has historically been either the actual number of tasks remaining in the DB, or the number of tasks in an in-memory buffer (current).

Either approach is an unreliable way to determine whether a TaskList is empty. Reading the number of tasks in the DB is unreliable when there's a large backlog of tasks due to needing to scan the partition. The in-memory buffer is inaccurate if persistence latency is sufficiently high compared to the task dispatch rate, and inaccurate when the TaskListManager hasn't loaded any tasks yet on startup.

Due to allocating TaskIDs in blocks, incrementing the block every time the TaskList lease changes, we can't use the offsets to determine the size of the backlog. We can however use the offsets to easily determine whether a TaskList is empty, comparing the ack level to the max read level(effectively the highest written task ID).

Add a new field `empty` to support this concept as part of the TaskListStatus.